Move package src/wallet/crypto to src/cipher/crypto
