+++ src/wallet/collection/wallet_test.go
@@ -7,8 +7,8 @@ import (
 	"testing"
 
 	"github.com/skycoin/skycoin/src/cipher"
+	"github.com/skycoin/skycoin/src/cipher/crypto"
 	"github.com/skycoin/skycoin/src/wallet"
-	"github.com/skycoin/skycoin/src/wallet/crypto"
 	"github.com/stretchr/testify/require"
 )
