+++ src/wallet/service.go
@@ -13,8 +13,8 @@ import (
 
 	"github.com/skycoin/skycoin/src/cipher"
 	"github.com/skycoin/skycoin/src/cipher/bip44"
+	"github.com/skycoin/skycoin/src/cipher/crypto"
 	"github.com/skycoin/skycoin/src/util/file"
-	"github.com/skycoin/skycoin/src/wallet/crypto"
 )
 
 // TransactionsFinder interface for finding address related transaction hashes
